+++ lib/evaluator.mli
@@ -1,0 +1,1 @@
+val run_exn : Raw_program.t -> Raw_term.t