+++ lib/raw_syntax/raw_program.ml
@@ -1,3 +1,5 @@
+[@@@coverage off]
+
 type t = rule list
 
 and rule = attr_list * Symbol.t * Symbol.t list * Raw_term.t